Moved goto definition to WR extension
