--- lib/Core/GotoDefinition/Exception/GotoDefinitionException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Phpactor\Core\GotoDefinition\Exception;
-
-class GotoDefinitionException extends \DomainException
-{
-}
