+++ lib/Extension/WorseReflection/GotoDefinition/Exception/GotoDefinitionException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Phpactor\Extension\WorseReflection\GotoDefinition\Exception;
+
+class GotoDefinitionException extends \DomainException
+{
+}
